+++ x-pack/plugins/security/public/management/roles/edit_role/privileges/es/index_privilege_form.test.tsx
@@ -107,70 +107,6 @@ test('should not render clusters field for local indices', () => {
   expect(wrapper.find('[data-test-subj="clustersInput0"]')).toHaveLength(0);
 });
 
-test('should render clusters field for remote indices', () => {
-  const wrapper = shallowWithIntl(
-    <IndexPrivilegeForm
-      indexType="remote_indices"
-      indexPrivilege={{
-        clusters: [],
-        names: [],
-        privileges: [],
-        query: '',
-        field_security: {
-          grant: [],
-        },
-      }}
-      remoteClusters={[
-        {
-          name: 'test1',
-          mode: 'proxy',
-          isConnected: false,
-          initialConnectTimeout: '30s',
-          skipUnavailable: false,
-          proxyAddress: 'localhost:9400',
-          proxySocketConnections: 18,
-          connectedSocketsCount: 0,
-          serverName: 'localhost',
-          securityModel: 'certificate',
-        },
-        {
-          name: 'test2',
-          mode: 'proxy',
-          isConnected: false,
-          initialConnectTimeout: '30s',
-          skipUnavailable: false,
-          proxyAddress: 'localhost:9400',
-          proxySocketConnections: 18,
-          connectedSocketsCount: 0,
-          serverName: 'localhost',
-          securityModel: 'api_key',
-        },
-      ]}
-      formIndex={0}
-      indexPatterns={[]}
-      indicesAPIClient={indicesAPIClientMock.create()}
-      availableIndexPrivileges={['all', 'read', 'write', 'index']}
-      isRoleReadOnly={false}
-      allowDocumentLevelSecurity
-      allowFieldLevelSecurity
-      validator={new RoleValidator()}
-      onChange={jest.fn()}
-      onDelete={jest.fn()}
-    />
-  );
-  const clustersInput = wrapper.find('[data-test-subj="clustersInput0"]');
-  expect(clustersInput).toHaveLength(1);
-  expect(clustersInput.prop('options')).toEqual([
-    { label: 'test2' },
-    { label: expect.anything(), isGroupLabelOption: true },
-    {
-      label: 'test1',
-      disabled: true,
-      append: expect.anything(),
-    },
-  ]);
-});
-
 describe('delete button', () => {
   const props = {
     indexType: 'indices' as const,
